removed cluster files. reworked position of some methods. added bad combinations of behavior params and pruning of this configs at the moment of simulation.
