--- requirements_cluster.txt
@@ -1,3 +1,0 @@
-numpy>=1.19.5
-pandas
-Pillow>=8.4.0
